+++ src/components/OrderPanel/BookingFixedDurationForm/BookingFixedDurationForm.js
@@ -12,6 +12,8 @@ import { Form, H6, PrimaryButton, FieldSelect } from '../../../components';
 import EstimatedCustomerBreakdownMaybe from '../EstimatedCustomerBreakdownMaybe';
 import FieldDateAndTimeInput from './FieldDateAndTimeInput';
 
+import FetchLineItemsError from '../FetchLineItemsError/FetchLineItemsError.js';
+
 import css from './BookingFixedDurationForm.module.css';
 
 // When the values of the form are updated we need to fetch
@@ -251,11 +253,7 @@ export const BookingFixedDurationForm = props => {
               </div>
             ) : null}
 
-            {fetchLineItemsError ? (
-              <span className={css.sideBarError}>
-                <FormattedMessage id="BookingFixedDurationForm.fetchLineItemsError" />
-              </span>
-            ) : null}
+            <FetchLineItemsError error={fetchLineItemsError} />
 
             <div className={css.submitButton}>
               <PrimaryButton