+++ src/components/OrderPanel/BookingTimeForm/BookingTimeForm.js
@@ -12,6 +12,8 @@ import { Form, H6, PrimaryButton, FieldSelect } from '../../../components';
 import EstimatedCustomerBreakdownMaybe from '../EstimatedCustomerBreakdownMaybe';
 import FieldDateAndTimeInput from './FieldDateAndTimeInput';
 
+import FetchLineItemsError from '../FetchLineItemsError/FetchLineItemsError.js';
+
 import css from './BookingTimeForm.module.css';
 
 // When the values of the form are updated we need to fetch
@@ -252,12 +254,7 @@ export const BookingTimeForm = props => {
                 />
               </div>
             ) : null}
-
-            {fetchLineItemsError ? (
-              <span className={css.sideBarError}>
-                <FormattedMessage id="BookingTimeForm.fetchLineItemsError" />
-              </span>
-            ) : null}
+            <FetchLineItemsError error={fetchLineItemsError} />
 
             <div className={css.submitButton}>
               <PrimaryButton